add build msvc16_sdl_x64 add StdBaseTsd.h fix Win64_SDL

+++ src/Plugins/SDLFileGroupPlugin/SDLFileInputStream.cpp
@@ -81,7 +81,7 @@ namespace Mengine
             return false;
         }
 
-        if( _offset + _size > size )
+        if( _offset + _size > (size_t)size )
         {
             LOGGER_ERROR( "invalid file '%s' range %d:%d size %d"
                 , fullPath
